+++ pytket/extensions/qiskit/tket_pass.py
@@ -78,6 +78,7 @@ class TketAutoPass(TketPass):
         self,
         backend: BackendV2,
         optimisation_level: int = 2,
+        instance: str | None = None,
         token: str | None = None,
     ):
         """Identifies a Qiskit backend and provides the corresponding default
@@ -89,10 +90,11 @@ class TketAutoPass(TketPass):
             compilation. Level 0 just solves the device constraints without
             optimising. Level 1 additionally performs some light optimisations.
             Level 2 adds more computationally intensive optimisations. Defaults to 2.
+        :param instance: Instance for the :py:class:`~qiskit_ibm_runtime.QiskitRuntimeService`.
         :param token: Authentication token to use the :py:class:`~qiskit_ibm_runtime.QiskitRuntimeService`.
         """
         if isinstance(backend, AerSimulator):
             tk_backend = self._aer_backend_map[backend.name]()
         else:
-            tk_backend = IBMQBackend(backend.name, token=token)
+            tk_backend = IBMQBackend(backend.name, instance=instance, token=token)
         super().__init__(tk_backend.default_compilation_pass(optimisation_level))