+++ docs/changelog.md
@@ -4,6 +4,24 @@
 
 # Changelog
 
+## 0.70.0 (July 2025)
+
+- Fix overwriting of user-supplied `service` parameter to `IBMQBackend`.
+
+## 0.69.0 (July 2025)
+
+- Migrate to upgraded IBM Quantum Platform.
+- Add `instance` parameter for `TketAutoPass`.
+- Update minimum qiskit-ibm-runtime version requirement to 0.40.1.
+
+## 0.68.3 (June 2025)
+
+- Restrict qiskit version to < 2.1.
+
+## 0.68.2 (June 2025)
+
+- Update minimum pytket version requirement to 2.6.0.
+
 ## 0.68.0 (May 2025)
 
 - Add {py:class}`~qiskit.transpiler.passes.VF2PostLayout` to default compilation.